+++ features/admin.actions.v1/api/create-action.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
+ * Copyright (c) 2024-2025, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -39,7 +39,8 @@ const httpClient: HttpClientInstance = AsgardeoSPAClient.getInstance()
  * @returns Promise containing the response.
  * @throws Throws an IdentityAppsApiException if the request fails.
  */
-const createAction = (actionType: string, actionBody: ActionInterface): Promise<ActionResponseInterface> => {
+const createAction = <T extends ActionInterface>(actionType: string, actionBody: T):
+    Promise<ActionResponseInterface> => {
 
     const requestConfig: RequestConfigInterface = {
         data: actionBody,