+++ features/admin.connections.v1/components/create/custom-authentication-create-wizard.tsx
@@ -113,10 +113,9 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
     subTitle,
     onWizardClose,
     onIDPCreate,
-    "data-componentid": _componentId = "application-creation-adapter"
+    "data-componentid": _componentId = "custom-authentication"
 }: CustomAuthenticationCreateWizardPropsInterface): ReactElement => {
     const wizardRef: MutableRefObject<any> = useRef(null);
-    const [ alert, setAlert, alertComponent ] = useWizardAlert();
 
     const { CUSTOM_AUTHENTICATION_CONSTANTS: CustomAuthConstants } = ConnectionUIConstants;
 
@@ -203,7 +202,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                 size="small"
                 color="grey"
                 name={ !showSecret ? "eye" : "eye slash" }
-                data-componentid={ `${_componentId}-authentication-property-secret1-view-button` }
+                data-componentid={ `${_componentId}-endpoint-authentication-property-secret-view-button` }
                 onClick={ onClick }
             />
         </InputAdornment>
@@ -215,6 +214,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
 
     /**
      * This method checks whether an error message is attached to a specific field.
+     *
      * @param errors - Errors object
      * @returns `true` if the field has an error, `false` otherwise.
      */
@@ -224,6 +224,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
 
     /**
      * This method handles endpoint authentication type dropdown changes.
+     *
      * @param event - event associated with the dropdown change.
      * @param data - data changed by the event
      */
@@ -233,6 +234,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
 
     /**
      * This method renders property fields of each endpoint authentication type.
+     *
      * @returns property fields of the selected authentication type.
      */
     const renderEndpointAuthPropertyFields = (): ReactElement => {
@@ -264,7 +266,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                             required={ true }
                             maxLength={ 100 }
                             minLength={ 0 }
-                            data-componentid={ `${_componentId}-authentication-property-username` }
+                            data-componentid={ `${_componentId}-endpoint-authentication-property-username` }
                             width={ 15 }
                         />
                         <Field.Input
@@ -289,7 +291,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                             required={ true }
                             maxLength={ 100 }
                             minLength={ 0 }
-                            data-componentid={ `${_componentId}-authentication-property-password` }
+                            data-componentid={ `${_componentId}-endpoint-authentication-property-password` }
                             width={ 15 }
                         />
                     </>
@@ -319,7 +321,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                             required={ true }
                             maxLength={ 100 }
                             minLength={ 0 }
-                            data-componentid={ `${_componentId}-authentication-property-accessToken` }
+                            data-componentid={ `${_componentId}-endpoint-authentication-property-accessToken` }
                             width={ 15 }
                         />
                     </>
@@ -344,7 +346,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                             required={ true }
                             maxLength={ 100 }
                             minLength={ 0 }
-                            data-componentid={ `${_componentId}-authentication-property-header` }
+                            data-componentid={ `${_componentId}-endpoint-authentication-property-header` }
                             width={ 15 }
                         />
                         <Field.Input
@@ -369,7 +371,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                             required={ true }
                             maxLength={ 100 }
                             minLength={ 0 }
-                            data-componentid={ `${_componentId}-authentication-property-value` }
+                            data-componentid={ `${_componentId}-endpoint-authentication-property-value` }
                             width={ 15 }
                         />
                     </>
@@ -381,6 +383,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
 
     /**
      * This method validates the endpoint configurations.
+     *
      * @param values - values to be validated.
      * @returns errors object.
      */
@@ -758,7 +761,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                         showTooltips={ true }
                         overlay={ renderDimmerOverlay() }
                         overlayOpacity={ 0.6 }
-                        data-componentid={ `${_componentId}-form-wizard-external-custom-authentication-
+                        data-componentid={ `${_componentId}-create-wizard-external-custom-authentication-
                         selection-card` }
                     />
                     <SelectionCard
@@ -799,7 +802,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                         contentTopBorder={ false }
                         overlay={ renderDimmerOverlay() }
                         overlayOpacity={ 0.6 }
-                        data-componentid={ `${_componentId}-form-wizard-internal-custom-authentication-
+                        data-componentid={ `${_componentId}-create-wizard-internal-user-custom-authentication-
                         selection-card` }
                     />
                     <SelectionCard
@@ -840,7 +843,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                         overlay={ renderDimmerOverlay() }
                         overlayOpacity={ 0.6 }
                         contentTopBorder={ false }
-                        data-componentid={ `${_componentId}-form-wizard-two-factor-custom-authentication-
+                        data-componentid={ `${_componentId}-create-wizard-two-factor-custom-authentication-
                         selection-card` }
                     />
                 </div>
@@ -881,7 +884,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                 required={ true }
                 maxLength={ 100 }
                 minLength={ 3 }
-                data-componentid={ `${_componentId}-form-wizard-identifier` }
+                data-componentid={ `${_componentId}-create-wizard-identifier` }
                 width={ 15 }
             />
             <Hint>{ t("customAuthentication:fields.createWizard.generalSettingsStep.identifier.hint") }</Hint>
@@ -895,7 +898,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                 required={ true }
                 maxLength={ 100 }
                 minLength={ 3 }
-                data-componentid={ `${_componentId}-form-wizard-display-name` }
+                data-componentid={ `${_componentId}-create-wizard-display-name` }
                 width={ 15 }
             />
         </WizardPage>
@@ -914,7 +917,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                 required={ true }
                 maxLength={ 100 }
                 minLength={ 0 }
-                data-componentid={ `${_componentId}-endpointUri` }
+                data-componentid={ `${_componentId}-create-wizard-endpoint-uri` }
                 width={ 15 }
             />
             <Divider className="divider-container" />
@@ -947,7 +950,7 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
                     ] }
                     onChange={ handleDropdownChange }
                     enableReinitialize={ true }
-                    data-componentid={ `${_componentId}-endpoint_authentication-dropdown` }
+                    data-componentid={ `${_componentId}-create-wizard-endpoint-authentication-dropdown` }
                     width={ 15 }
                 />
                 <div className="box-field">{ renderEndpointAuthPropertyFields() }</div>
@@ -1025,7 +1028,9 @@ const CustomAuthenticationCreateWizard: FunctionComponent<CustomAuthenticationCr
             data-componentid={ `${_componentId}-modal` }
         >
             <ModalWithSidePanel.MainPanel>
-                <ModalWithSidePanel.Header className="wizard-header" data-componentid={ `${_componentId}-modal-header` }>
+                <ModalWithSidePanel.Header
+                    className="wizard-header"
+                    data-componentid={ `${_componentId}-modal-header` }>
                     <div className={ "display-flex" }>
                         <GenericIcon
                             icon={ ConnectionsManagementUtils.resolveConnectionResourcePath(