+++ features/admin.policy-administration.v1/components/policy-list-node.tsx
@@ -22,7 +22,6 @@ import Typography from "@oxygen-ui/react/Typography";
 import { AppConstants, history } from "@wso2is/admin.core.v1";
 import { AlertLevels, IdentifiableComponentInterface } from "@wso2is/core/models";
 import { addAlert } from "@wso2is/core/store";
-import kebabCase from "lodash-es/kebabCase";
 import React, { FunctionComponent, HTMLAttributes, ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
@@ -52,26 +51,24 @@ export interface PolicyListDraggableNodePropsInterface
 
 const PolicyListNode: FunctionComponent<PolicyListDraggableNodePropsInterface> = ({
     "data-componentid": componentId = "policy-list--node",
-    id,
     policy,
     mutateInactivePolicyList,
     setInactivePolicies,
     setPageInactive,
     setHasMoreInactivePolicies,
-    mutateActivePolicyList,
-    ...rest
+    mutateActivePolicyList
 }: PolicyListDraggableNodePropsInterface): ReactElement => {
     const { t } = useTranslation();
 
     const dispatch: Dispatch = useDispatch();
 
     const handleEdit = (policyId: string) => {
-        history.push(`${AppConstants.getPaths().get("EDIT_POLICY").replace(":id", kebabCase(policyId))}`);
+        history.push(`${AppConstants.getPaths().get("EDIT_POLICY").replace(":id", btoa(policyId))}`);
     };
 
     const handleDelete = async (): Promise<void> => {
         try {
-            await deletePolicy(policy.policyId);
+            await deletePolicy(btoa(policy.policyId));
 
             setPageInactive(0);
             setHasMoreInactivePolicies(true);
@@ -115,10 +112,6 @@ const PolicyListNode: FunctionComponent<PolicyListDraggableNodePropsInterface> =
                 message: t("policyAdministration:alerts.activateSuccess.message")
             }));
 
-            setPageInactive(0);
-            setHasMoreInactivePolicies(true);
-            setInactivePolicies([]);
-
             mutateActivePolicyList();
             mutateInactivePolicyList();
 
@@ -141,17 +134,17 @@ const PolicyListNode: FunctionComponent<PolicyListDraggableNodePropsInterface> =
                                 <Icon
                                     onClick={ handleActivate }
                                     data-componentid={ `${componentId}-edit-button` }
-                                    className="list-icon"
+                                    className="policy-list-icon list-icon"
                                     size="massive"
                                     color="grey"
                                     name="chevron left"
                                 />
                             ) }
                             position="top center"
-                            content={ "Deactivate" }
+                            content={ t("policyAdministration:popup.deactivate") }
                             inverted
                         />
-                        <Typography>{ policy.policyId }</Typography>
+                        <Typography className="ellipsis-text">{ policy.policyId }</Typography>
                     </Stack>
                     <Stack direction={ "row" } marginTop={ "3px" }>
                         <Popup
@@ -167,7 +160,7 @@ const PolicyListNode: FunctionComponent<PolicyListDraggableNodePropsInterface> =
                                 />
                             ) }
                             position="top center"
-                            content={ "Edit" }
+                            content={ t("common:edit") }
                             inverted
                         />
                         <Popup
@@ -182,7 +175,7 @@ const PolicyListNode: FunctionComponent<PolicyListDraggableNodePropsInterface> =
                                 />
                             ) }
                             position="top center"
-                            content={ "Delete" }
+                            content={ t("common:delete") }
                             inverted
                         />
                     </Stack>