+++ features/admin.users.v1/components/edit-user.tsx
@@ -111,6 +111,7 @@ export const EditUser: FunctionComponent<EditUserPropsInterface> = (
     const userRolesDisabledFeatures: string[] = useSelector((state: AppState) => {
         return state.config.ui.features?.users?.disabledFeatures;
     });
+    const userSchemaURI: string = useSelector((state: AppState) => state?.config?.ui?.userSchemaURI);
 
     const isUpdatingSharedProfilesEnabled: boolean = !userRolesDisabledFeatures?.includes(
         UserManagementConstants.FEATURE_DICTIONARY.get("USER_SHARED_PROFILES")
@@ -124,8 +125,8 @@ export const EditUser: FunctionComponent<EditUserPropsInterface> = (
         if (!isFeatureEnabled(featureConfig?.users, UserManagementConstants.FEATURE_DICTIONARY.get("USER_UPDATE"))
             || readOnlyUserStores?.includes(userStore?.toString())
             || !hasUsersUpdatePermissions
-            || user[ SCIMConfigs.scim.enterpriseSchema ]?.userSourceId
-            || user[ UserManagementConstants.CUSTOMSCHEMA ]?.isReadOnlyUser === "true"
+            || user[ SCIMConfigs.scim.systemSchema ]?.userSourceId
+            || user[ userSchemaURI ]?.isReadOnlyUser === "true"
         ) {
             setReadOnly(true);
         }
@@ -141,7 +142,7 @@ export const EditUser: FunctionComponent<EditUserPropsInterface> = (
     }, [ user ]);
 
     useEffect(() => {
-        if (user[ SCIMConfigs.scim.enterpriseSchema ]?.managedOrg) {
+        if (user[ SCIMConfigs.scim.systemSchema ]?.managedOrg) {
             if (!isUpdatingSharedProfilesEnabled) {
                 setIsUserProfileReadOnly(true);
             }
@@ -225,7 +226,7 @@ export const EditUser: FunctionComponent<EditUserPropsInterface> = (
                         isUserManagedByParentOrg={ isUserManagedByParentOrg }
                         adminUserType={ AdminAccountTypes.INTERNAL }
                         allowDeleteOnly={
-                            user[ UserManagementConstants.CUSTOMSCHEMA ]?.isReadOnlyUser === "true"
+                            user[ userSchemaURI ]?.isReadOnlyUser === "true"
                         }
                         editUserDisclaimerMessage={ (
                             <Grid>