+++ modules/access-control/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @wso2is/access-control
 
+## 3.3.12
+
+### Patch Changes
+
+- Updated dependencies [[`31b4e92ae73fea1372a03f9aa0511f062030e44f`](https://github.com/wso2/identity-apps/commit/31b4e92ae73fea1372a03f9aa0511f062030e44f)]:
+  - @wso2is/core@2.5.8
+
 ## 3.3.11
 
 ### Patch Changes